career-advisor/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.15
pydantic==1.10.7
openai==0.27.4
scikit-learn==1.2.2
pandas==2.0.2
numpy==1.24.3
python-dotenv==1.0.0
psycopg2-binary==2.9.6
pytest==7.3.1
httpx==0.24.1
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0 